+++ package.json
@@ -1,5 +1,5 @@
 {
-    "name": "nativescript-urban-airship",
+    "name": "nativescript-ns-urbanairship",
     "version": "1.0.0",
     "description": "Your awesome NativeScript plugin.",
     "main": "urbanairship",
